Refactor MyAdminProfile, UserManagement
