+++ src/components/Admin/UsersManagement/BannedWidget/index.js
