+++ src/pages/admin/users/sensei-profile/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Helmet } from 'react-helmet'
-import SenseiProfileComponent from '../../../components/Admin/UsersManagement/SenseiProfileComponent'
+import SenseiProfileComponent from '../../../../components/Admin/UsersManagement/SenseiProfileComponent'
 
 const UserProfile = () => {
   return (
