+++ src/pages/admin/users/student-profile/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Helmet } from 'react-helmet'
-import StudentProfileComponent from '../../../components/Admin/UsersManagement/StudentProfileComponent'
+import StudentProfileComponent from '../../../../components/Admin/UsersManagement/StudentProfileComponent'
 
 const UserProfile = () => {
   return (
